fixed memory issue + small optimization in obstacle gen (see desc)
running physics.clear() messed with how i was already clearing physics objects--it created a sort of memory "leak" where there were a bunch of references created to stuff that never got used again and kept building up. but i actually realized that clearing all physics objects and just re-adding the player is way more efficient than what I was doing before.

however, clearing all physics objects and re-adding to the world doesn't clear the objects from the worldnode :) so removePolyNodes is still a needed function.

although, come to think of it. that just gave me an idea for something. if there's a commit after that that means the idea worked