trigger messages!
Note: the text does not currently change as I am not sure of the path to the label object and Czar has not responded to me on slack. This message is so long now. I am happy this works but I wanted to finish it finish it today. Oh well I guess. At least we know that it works.